refactor(rees): promote reconstructOldContent into a shared analyzer helper
reconstructOldContent (unified-diff reverse-patch reconstruction) was
private to doc-comment-drift.ts and imported cross-file from there by
exhaustiveness-drift.ts -- a one-off trick, not shared infrastructure.
Move it into its own co-located module (matching the existing
diff-lines.ts / binary-extensions.ts / github-headers.ts precedent for
shared analyzer helpers) so any analyzer can recover a changed file's
pre-PR text without re-deriving this.

Pure code motion: every executable line is byte-identical (verified via
diff + matching MD5 after stripping whole-line comments); only the
function's own doc comment and one inline comment were reworded to
drop doc-comment-drift-specific framing now that the helper is shared.
Both existing callers are migrated with a one-line import-path change
each and no other modifications.

Adds a dedicated test file for the promoted helper: the four existing
reconstructOldContent unit tests move over verbatim, plus five new
tests closing every previously-untested branch (a non-hunk preamble
line, out-of-order/overlapping hunks, a "no newline at end of file"
marker, the trailing-flush no-op case, and the wholly-new-file case) --
100% line/branch/function coverage confirmed via node's built-in
coverage instrumentation.
